Complicated route to fix
Merged the DB stuff I did in preparation for KSP 1.2 (which was done due
to the change in OnSave, OnLoad order between Scenario and Part). This
will have to be revised as soon as KSP1.2 is completed. NO MORE WORK TO
BE DONE UNTIL 1.2 IS FULLY RELEASED FOR SAKE OF CLEANLINESS (refactoring
as well).

some null reference checks are made now before saving/loading; this
should prevent the game breaking issue of not being able to save/change
scenes if RoverScience part is not present